python/requirements.txt
ddginternal
groq
fastapi
uvicorn
python-dotenv